Add workaround for build-time disk space issue
The one-liner added in this commit was suggested by Et7f3 in the Reason
community. It cleans out build artifacts when the dependency is done
building. This should ensure that disk space limits are not hit.

We've slightly adjusted it to keep the script running until compilation
is complete. The original script ran indefinitely but this would cause
the Platform.sh build not to finish. We use the `&` marker to detach the
process so that we can continue doing the actual build in the meantime.