Make esy cache writeable during runtime
Running the executable through `esy` is needed to ensure the path
contains all the necessary libraries. However, it appears that starting
`esy` (even just `esy -v`) opens the esy cache in write mode. This is
not allowed on a read-only system.

This commit tries to keep the change from read-only to writeable as
small as possible (as is good practice).