Dump app environment and load without esy
The cache is not enough writeable paths for esy. However, we don't want
to just open up the entire filesystem to esy after deployment.

Et7f3 and ulrikstrid suggested a way of dumping the shell environment
from esy to disk during build. This can then be used after deployment to
ensure all libraries can be found.

This allows us to remove the mount we created earlier.