Limit the environment variables stored during build
The previous command dumped the entire shell environment which included
PLATFORM_ variables that change between build and deployment (and
shouldn't be written to disk). It also dumped the IFS variable which was
multiline and broke with our `xargs` command.

To solve both of these issues we only dump the variables we think we
need.