Implement killswitch for cleanup process
We currently have some link errors which cause the background process to
keep running. This means we need to wait for the build to timeout.

If we instead create our executable also when that fails, the build can
error out earlier.

This can be removed when we have a working linker process.