Use Platform.sh build cache
I mixed up how the platform cache works with mounts, so no caching was
actually happening. We configure esy to use the Platform.sh build cache
for its storage (docs at
https://esy.sh/docs/en/esy-configuration.html#esy-prefix-path ). This
works because we don't need esy after building with our `npm pack`
strategy.

We also adjust the background clean-up script since the build-cache is
limited to 4 GB in storage space (but in return it does not affect the
space used by the application during runtime).